+++ src/containers/Header/ApplicationScreenTitle/ApplicationScreenTitle.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useMemo } from 'react';
-import { useSelector } from 'react-redux';
 import { useTranslation } from 'react-i18next';
+import { useAppSelector } from 'asab_webui_components';
 
 import './ApplicationScreenTitle.scss';
 
@@ -16,17 +16,22 @@ import './ApplicationScreenTitle.scss';
 		hasHeaderTitle: false,
 	}
 
-	ReactDOM.render((
+	root.render(
 		<HashRouter>
-			<Application modules={modules} configDefaults={configDefaults}/>
+			<Application
+				configdefaults={ConfigDefaults} 
+				modules={modules}
+			/>
 		</HashRouter>
-	), document.getElementById('app'));
+	);
 */
 
 const ApplicationScreenTitle = (props) => {
-	const subtitle = useSelector(state => state.header?.subtitle);
-	const title = useSelector(state => state.config?.title);
-	const hasHeaderTitle = useSelector(state => state?.config?.hasHeaderTitle);
+
+	const subtitle = useAppSelector(state => state.header?.subtitle);
+	const title = useAppSelector(state => state.config?.title);
+	const hasHeaderTitle = useAppSelector(state => state?.config?.hasHeaderTitle);
+
 	const { t } = useTranslation();
 	const route = props.routes?.[0];