+++ src/index.js
@@ -5,5 +5,4 @@ export { getBrandImage } from './components/branding/BrandImage';
 export { default as I18nModule } from './modules/i18n';
 export { default as TenantModule } from './modules/tenant';
 export { default as AuthModule } from './modules/auth';
-export { default as AuthReducer } from './modules/auth/reducer.js'; // Export auth reducer to be used in Login
 export { default as AboutModule } from './modules/about';