+++ CHANGELOG.md
@@ -1,8 +1,28 @@
 # CHANGELOG for ASAB WebUI Shell
 
+## 26.1.3
+
+- Refactor Application store with createStore() from asab_webui_components to make global variables dynamic and fix issues with not updating the AppStore state dynamically. (#31)
+
+## 26.1.2
+
+- Import FlowbiteIllustration from `asab_webui_components` into informational screens (#27)
+
+## 26.1.1
+
+- Upgrade to react v19, remove redux dependency and replace it with custom redux-like context and AppStore, remove unused flag header (#17)
+
+## 25.2.8
+
+- Update page orientation in the theme-print styles (#21)
+
+## 25.2.7
+
+- Implement refresh token session validation to the Auth module (#22)
+
 ## 25.2.6
 
-- Updated InvitationScreen. Update text and add margin for input (!18)
+- Updated InvitationScreen. Update text and add margin for input (#18)
 
 ## 25.2.5